requirements.txt
bcrypt==4.1.2
colorama==0.4.6
psycopg2==2.9.9
python-dotenv==1.0.1
SQLAlchemy==2.0.29
typing_extensions==4.11.0
